requirements.txt
langchain
langchain-community
langchain-ollama
langchain-openai
langchain-qdrant
qdrant-client
redis
fastapi
uvicorn
streamlit
httpx
beautifulsoup4
tiktoken
pydantic
python-dotenv
sentence-transformers
arxiv
pandas
sqlalchemy
langgraph
tavily-python
google-search-results
altair<5
